Interfaz grafica totalmente finalizada y funcional
Falta la implementacion de la base de datos. Revisar el TODO de estudiante y persona respecto a la ID

+++ src/Estudiante.cs
@@ -9,7 +9,7 @@ public class Estudiante: Persona
     public bool activo { get{return _activo;} set{ _activo = value; }}
     public double promedio { get{return _promedio;} set{ _promedio = value; }}
 
-    public Estudiante( double ID, string nombres, string  apellidos, string email, double codigo, Programa programa, bool activo, double promedio): base(ID, nombres, apellidos, email) {
+    public Estudiante(int ID, string nombres, string  apellidos, string email, double codigo, Programa programa, bool activo, double promedio): base(ID, nombres, apellidos, email) {
         this._codigo = codigo;
         this._programa = programa;
         this._activo = activo;
